+++ packages/experience-builder-sdk/src/blocks/preview/CompositionBlock.tsx
@@ -73,11 +73,6 @@ export const CompositionBlock = ({
   }, [isAssembly, node.definitionId]);
 
   const nodeProps = useMemo(() => {
-    // Don't enrich the assembly wrapper node with props
-    if (!componentRegistration || isAssembly) {
-      return {};
-    }
-
     const propMap: Record<string, PrimitiveValue> = {
       cfSsrClassName: node.variables.cfSsrClassName
         ? resolveDesignValue(
@@ -86,11 +81,16 @@ export const CompositionBlock = ({
           )
         : undefined,
     };
+    // Don't enrich the assembly wrapper node with props
+    if (!componentRegistration || isAssembly) {
+      return propMap;
+    }
 
-    return Object.entries(componentRegistration.definition.variables).reduce(
+    const props = Object.entries(componentRegistration.definition.variables).reduce(
       (acc, [variableName, variableDefinition]) => {
         const variable = node.variables[variableName];
         if (!variable) return acc;
+
         switch (variable.type) {
           case 'DesignValue':
             acc[variableName] = resolveDesignValue(variable.valuesByBreakpoint, variableName);
@@ -145,9 +145,29 @@ export const CompositionBlock = ({
       },
       propMap,
     );
+
+    if (componentRegistration.definition.slots) {
+      for (const slotId in componentRegistration.definition.slots) {
+        const slotNode = node.children.find((child) => child.slotId === slotId);
+        if (slotNode) {
+          props[slotId] = (
+            <CompositionBlock
+              node={slotNode}
+              locale={locale}
+              hyperlinkPattern={hyperlinkPattern}
+              entityStore={entityStore}
+              resolveDesignValue={resolveDesignValue}
+            />
+          );
+        }
+      }
+    }
+
+    return props;
   }, [
     componentRegistration,
     isAssembly,
+    node.children,
     node.variables,
     resolveDesignValue,
     entityStore,