+++ packages/visual-editor/src/components/Dropzone/useComponentProps.ts
@@ -3,7 +3,7 @@ import {
   buildCfStyles,
   calculateNodeDefaultHeight,
   isLinkToAsset,
-  isEmptyStructureWithRelativeHeight,
+  isStructureWithRelativeHeight,
   isContentfulStructureComponent,
   transformBoundContentValue,
   resolveHyperlinkPattern,
@@ -47,9 +47,10 @@ type UseComponentProps = {
   node: ExperienceTreeNode;
   resolveDesignValue: ResolveDesignValueType;
   areEntitiesFetched: boolean;
-  definition: ComponentRegistration['definition'];
+  definition?: ComponentRegistration['definition'];
   renderDropzone: RenderDropzoneFunction;
   userIsDragging: boolean;
+  slotId?: string;
 };
 
 export const useComponentProps = ({
@@ -59,12 +60,18 @@ export const useComponentProps = ({
   renderDropzone,
   definition,
   userIsDragging,
+  slotId,
 }: UseComponentProps) => {
   const unboundValues = useEditorStore((state) => state.unboundValues);
   const hyperlinkPattern = useEditorStore((state) => state.hyperLinkPattern);
   const locale = useEditorStore((state) => state.locale);
   const dataSource = useEditorStore((state) => state.dataSource);
   const entityStore = useEntityStore((state) => state.entityStore);
+
+  const isEmptyZone = useMemo(() => {
+    return !node.children.filter((child) => child.data.slotId === slotId).length;
+  }, [node.children, slotId]);
+
   const props: ComponentProps = useMemo(() => {
     const propsBase = {
       cfSsrClassName: node.data.props.cfSsrClassName
@@ -183,9 +190,19 @@ export const useComponentProps = ({
       {},
     );
 
+    const slotProps: Record<string, React.JSX.Element> = {};
+    if (definition.slots) {
+      for (const slotId in definition.slots) {
+        slotProps[slotId] = renderDropzone(node, {
+          zoneId: [node.data.id, slotId].join('|'),
+        });
+      }
+    }
+
     return {
       ...propsBase,
       ...extractedProps,
+      ...slotProps,
     };
   }, [
     hyperlinkPattern,
@@ -197,9 +214,10 @@ export const useComponentProps = ({
     areEntitiesFetched,
     unboundValues,
     entityStore,
+    renderDropzone,
   ]);
 
-  const cfStyles = buildCfStyles(props as StyleProps, node.data.blockId);
+  const cfStyles = buildCfStyles(props as StyleProps);
 
   // Separate the component styles from the editor wrapper styles
   const { margin, height, width, maxWidth, ...componentStyles } = cfStyles;
@@ -231,9 +249,10 @@ export const useComponentProps = ({
       width: '100%',
       height: '100%',
       maxWidth: 'none',
-      ...(isEmptyStructureWithRelativeHeight(node.children.length, node?.data.blockId, height) && {
-        minHeight: EMPTY_CONTAINER_HEIGHT,
-      }),
+      ...(isEmptyZone &&
+        isStructureWithRelativeHeight(node?.data.blockId, height) && {
+          minHeight: EMPTY_CONTAINER_HEIGHT,
+        }),
       ...(userIsDragging &&
         isContentfulStructureComponent(node?.data.blockId) &&
         node?.data.blockId !== CONTENTFUL_COMPONENTS.columns.id && {
@@ -260,10 +279,13 @@ export const useComponentProps = ({
     node,
     renderDropzone,
     ...omit(props, stylesToRemove, ['cfHyperlink', 'cfOpenInNewTab', 'cfSsrClassName']),
-    ...(definition.children ? { children: renderDropzone(node) } : {}),
+    ...(definition?.children ? { children: renderDropzone(node) } : {}),
   };
 
-  return { componentProps, wrapperProps };
+  return {
+    componentProps,
+    wrapperProps,
+  };
 };
 
 const addExtraDropzonePadding = (padding: string) =>