+++ packages/visual-editor/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/contentful/experience-builder/compare/v1.7.1-beta.0...v1.7.1) (2024-06-11)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.7.0](https://github.com/contentful/experience-builder/compare/v1.7.0-beta.0...v1.7.0) (2024-06-07)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.6.0](https://github.com/contentful/experience-builder/compare/v1.6.0-beta.0...v1.6.0) (2024-06-03)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.5.2](https://github.com/contentful/experience-builder/compare/v1.5.2-beta.0...v1.5.2) (2024-05-30)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.5.1](https://github.com/contentful/experience-builder/compare/v1.5.1-beta.2...v1.5.1) (2024-05-27)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.5.1-beta.2](https://github.com/contentful/experience-builder/compare/v1.5.1-beta.1...v1.5.1-beta.2) (2024-05-27)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.5.1-beta.1](https://github.com/contentful/experience-builder/compare/v1.5.1-beta.0...v1.5.1-beta.1) (2024-05-27)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
+## [1.5.1-beta.0](https://github.com/contentful/experience-builder/compare/v1.5.0...v1.5.1-beta.0) (2024-05-27)
+
+**Note:** Version bump only for package @contentful/experiences-visual-editor-react
+
 # [1.5.0](https://github.com/contentful/experience-builder/compare/v1.5.0-beta.0...v1.5.0) (2024-05-16)
 
 **Note:** Version bump only for package @contentful/experiences-visual-editor-react